Refactor code and add env configuration
