+++ .gitignore
@@ -6,3 +6,4 @@ __pycache__/
 .vscode/
 Backend/.env
 media/
+frontend/.env
