+++ Backend/.env
@@ -5,3 +5,5 @@ DB_USER=root
 DB_PASSWORD=12345
 DB_HOST=127.0.0.1
 DB_PORT=3306
+
+SERVER_IP=172.31.250.253
