+++ Backend/muebleria/settings.py
@@ -11,6 +11,7 @@ https://docs.djangoproject.com/en/5.2/ref/settings/
 """
 
 from pathlib import Path
+from decouple import config
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
@@ -20,12 +21,14 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # See https://docs.djangoproject.com/en/5.2/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = 'django-insecure-+-l2%56x%3e-5@th3*1@0d%vz#rnep@!e0ozjf-z4i89=c5!7g'
+SECRET_KEY = config('SECRET_KEY')
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
+DEBUG = config('DEBUG', default='False') == 'True'
 
-ALLOWED_HOSTS = ['192.168.1.52', 'localhost']
+SERVER_IP = config('SERVER_IP', default='127.0.0.1')
+
+ALLOWED_HOSTS = ['localhost', '127.0.0.1', SERVER_IP]
 
 
 
@@ -108,11 +111,9 @@ SIMPLE_JWT = {
     'REFRESH_TOKEN_LIFETIME': timedelta(days=1),
 }
 
-ALLOWED_HOSTS = ['localhost', '127.0.0.1', '192.168.1.52']
-
 CORS_ALLOWED_ORIGINS = [
     "http://localhost:5173",
-    "http://192.168.1.52:5173",
+    f"http://{SERVER_IP}:5173",
 ]
