+++ frontend/src/pages/FinancialReports.jsx
