+++ frontend/src/pages/InventoryReports.jsx
