+++ frontend/src/pages/SalesReports.jsx
